refactor(suggestions): split the read path by mode
Pure relocation, no logic touched. `bigram/query.rs` was at 105 lines and
`bigram/` was at the five-file limit, so the prediction mode had nowhere to go.

The two modes are also different questions. Reranking starts from a set the
prefix already produced and decides the order; predicting starts from nothing and
decides whether to speak at all. Splitting now keeps the second from being read
as a variation of the first.
