+++ crates/funput-suggestions/src/bigram/query/predict.rs
@@ -1,0 +1,3 @@
+//! Deciding whether the context is worth speaking about on its own.
+//!
+//! Nothing here yet; the thresholds arrive with the prediction mode.
