+++ crates/funput-suggestions/src/bigram/query/rerank.rs
@@ -1,4 +1,4 @@
-//! The read path: letting what came before reorder what comes next.
+//! Letting what came before reorder what comes next.
 //!
 //! A trie node remembers only its best few words by frequency, so a pair can be
 //! sharp — "xin" is almost always followed by "chào" — while its target is rare
@@ -6,29 +6,12 @@
 //! to *add* a candidate, not only to move one. Everything it adds still matches
 //! the prefix, so it is still a completion of what the user is typing.
 
-use super::follower::{FOLLOWER_SLOTS, Follower};
+use crate::bigram::follower::{FOLLOWER_SLOTS, Follower};
 use crate::engine::SuggestionEngine;
 use crate::index::{NONE, TOP_K, normalize};
-use crate::types::SuggestionSet;
 
 impl SuggestionEngine {
-    /// The words `prefix` could become, with the ones that have followed
-    /// `previous` before moved to the front.
-    ///
-    /// `previous` is the caller's business, as it is for `learn_after`: a
-    /// platform that cannot vouch for what came before passes `None` and gets
-    /// exactly what `suggest` has always returned.
-    #[inline]
-    pub fn suggest_with(&self, previous: Option<&str>, prefix: &str) -> SuggestionSet<'_> {
-        let (ids, len) = self.prefix_candidates(prefix);
-        let Some((context, _)) = previous.and_then(|text| self.context_slot(text)) else {
-            return self.assemble(ids, len);
-        };
-        let (ids, len) = self.with_context(context, prefix, ids, len);
-        self.assemble(ids, len)
-    }
-
-    fn with_context(
+    pub(super) fn with_context(
         &self,
         context: u32,
         prefix: &str,
